+++ src/Feed/OutputModel/Calendar/Resource.php
@@ -8,6 +8,7 @@ class Resource
 {
     public function __construct(
         public string $id,
+        public string $name,
         public string $locationId,
         public string $displayName,
     ) {}